+++ Libplanet/Action/ActionContext.cs
@@ -1,4 +1,4 @@
-using System;
+using System.Collections.Immutable;
 using System.Diagnostics.Contracts;
 using System.Security.Cryptography;
 using Libplanet.Assets;
@@ -12,7 +12,6 @@ namespace Libplanet.Action
     {
         private readonly int _randomSeed;
         private readonly ITrie? _previousBlockStatesTrie;
-        private readonly Predicate<Currency>? _nativeTokenPredicate;
         private HashDigest<SHA256>? _previousStateRootHash;
 
         public ActionContext(
@@ -27,7 +26,7 @@ namespace Libplanet.Action
             bool rehearsal = false,
             ITrie? previousBlockStatesTrie = null,
             bool blockAction = false,
-            Predicate<Currency>? nativeTokenPredicate = null)
+            IImmutableSet<Currency>? nativeTokens = null)
         {
             GenesisHash = genesisHash;
             Signer = signer;
@@ -41,7 +40,7 @@ namespace Libplanet.Action
             _randomSeed = randomSeed;
             _previousBlockStatesTrie = previousBlockStatesTrie;
             BlockAction = blockAction;
-            _nativeTokenPredicate = nativeTokenPredicate;
+            NativeTokens = nativeTokens ?? ImmutableHashSet<Currency>.Empty;
         }
 
         public BlockHash? GenesisHash { get; }
@@ -75,8 +74,7 @@ namespace Libplanet.Action
 
         public bool BlockAction { get; }
 
-        public bool IsNativeToken(Currency currency) =>
-            _nativeTokenPredicate is { } && _nativeTokenPredicate(currency);
+        public IImmutableSet<Currency> NativeTokens { get; }
 
         [Pure]
         public IActionContext GetUnconsumedContext() =>
@@ -92,6 +90,6 @@ namespace Libplanet.Action
                 Rehearsal,
                 _previousBlockStatesTrie,
                 BlockAction,
-                _nativeTokenPredicate);
+                NativeTokens);
     }
 }